Add data validation to stock inputs and other misc quality of life changes

+++ api/Controllers/StockController.cs
@@ -26,16 +26,23 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid) 
+                return BadRequest(ModelState); // Data Validation
+                
             var stockModel = stockDto.ToStockFromCreateDTO();
+            
             await _stockRepo.CreateAsync(stockModel);
 
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid) 
+                return BadRequest(ModelState); // Data Validation
+                
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null) 
@@ -50,6 +57,9 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) 
+                return BadRequest(ModelState); // Data Validation
+                
             var stocks = await _stockRepo.GetAllAsync();
 
             var stockDto = stocks.Select(s => s.ToStockDto()); // 'Select' => map
@@ -57,9 +67,13 @@ namespace api.Controllers
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")] // '{id}' => Model binding
+        [HttpGet] 
+        [Route("{id:int}")] // '{id}' => Model binding
         public async Task<IActionResult> GetById([FromRoute] int id) 
         {
+            if(!ModelState.IsValid) 
+                return BadRequest(ModelState); // Data Validation
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if(stock == null) 
@@ -71,9 +85,12 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto) 
         {
+            if(!ModelState.IsValid) 
+                return BadRequest(ModelState); // Data Validation
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if(stockModel == null)
